Unify 'MetadataSections' data structure, as well as common ELF and COFF procedures
