+++ stdlib/public/runtime/ImageInspectionCommon.h
@@ -1,4 +1,4 @@
-//===--- ImageInspectionELF.h -----------------------------------*- C++ -*-===//
+//===--- ImageInspectionCommon.h - Image inspection routines -----*- C++ -*-===//
 //
 // This source file is part of the Swift.org open source project
 //
@@ -12,34 +12,36 @@
 ///
 /// \file
 ///
-/// ELF specific image inspection routines.
+/// This file unifies common ELF and COFF image inspection routines
 ///
 //===----------------------------------------------------------------------===//
 
-#ifndef SWIFT_RUNTIME_IMAGEINSPECTIONELF_H
-#define SWIFT_RUNTIME_IMAGEINSPECTIONELF_H
+#ifndef SWIFT_RUNTIME_IMAGEINSPECTIONCOMMON_H
+#define SWIFT_RUNTIME_IMAGEINSPECTIONCOMMON_H
 
-#define SWIFT_REFLECTION_METADATA_ELF_NOTE_MAGIC_STRING "swift_reflection_metadata_magic_string"
+#if !defined(__MACH__)
 
 #if defined(__ELF__)
+#define SWIFT_REFLECTION_METADATA_ELF_NOTE_MAGIC_STRING "swift_reflection_metadata_magic_string"
+#endif // defined(__ELF__)
 
 #include "../SwiftShims/Visibility.h"
 #include <cstdint>
-#include <cstddef>
 
 namespace swift {
+struct MetadataSections;
+static constexpr const uintptr_t CurrentSectionMetadataVersion = 1;
+}
+
 struct SectionInfo {
   uint64_t size;
   const char *data;
 };
 
-static constexpr const uintptr_t CurrentSectionMetadataVersion = 1;
-} // namespace swift
-
 // Called by injected constructors when a dynamic library is loaded.
 SWIFT_RUNTIME_EXPORT
 void swift_addNewDSOImage(const void *addr);
 
-#endif // defined(__ELF__)
+#endif // !defined(__MACH__)
 
-#endif // SWIFT_RUNTIME_IMAGE_INSPECTION_ELF_H
+#endif // SWIFT_RUNTIME_IMAGEINSPECTIONCOMMON_H
