+++ stdlib/public/runtime/SwiftRT-ELF.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "ImageInspectionELF.h"
+#include "ImageInspectionCommon.h"
 #include "../SwiftShims/MetadataSections.h"
 
 #include <cstddef>
